feat: meme types
