+++ src/types/Meme.ts
@@ -1,0 +1,11 @@
+export interface Meme {
+  postLink: string;
+  subreddit: string;
+  title: string;
+  url: string;
+  nsfw: boolean;
+  spoiler: boolean;
+  author: string;
+  ups: number;
+  preview: string[];
+}
